Add give_rating form

+++ ratings/models.py
@@ -8,7 +8,7 @@ from django.db import models
 """
 class Passenger( models.Model ):
 	name = models.CharField( max_length=30)
-	rating = models.DecimalField( max_digits=2, decimal_places=1,default = 3.5) 
+	rating = models.FloatField( default = 3.5) 
 
 	def __str__( self ):
 		return self.name + "_" + str(self.rating)
@@ -22,8 +22,8 @@ class Passenger( models.Model ):
 """
 class Driver( models.Model ):
 	name = models.CharField( max_length= 30 )
-	rating = models.DecimalField( max_digits = 2, decimal_places = 1, default = 3.5 )
-	trips = models.DecimalField( max_digits = 10, decimal_places = 0, default = 0 )
+	rating = models.FloatField( default = 3.5 )
+	trips = models.IntegerField( default= 0 )
 	
 	def __str__( self ):
 		return self.name + "_" + str(self.rating)
